requirements.txt
flask
click==8.1.3
colorama==0.4.5
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-decouple==3.6
Werkzeug==2.1.2
